+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+It's now finally possible to add programs to startup
+
+### Added
+
+- #1 Added a startup command
+
 ## [1.2.1] - 2019-11-02
 
 No new application features, but streamlined build, publish and release process.
@@ -49,7 +55,8 @@ No new application features, but streamlined build, publish and release process.
 - Enable/Disable command
   - It's now possible to enable/disable existing programs starting with windows.
 
-[Unreleased]: https://github.com/Faustvii/StartupManager/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/Faustvii/StartupManager/compare/v1.2.1...HEAD
+[1.2.1]: https://github.com/Faustvii/StartupManager/compare/v1.2.0...v1.2.1
 [1.2.0]: https://github.com/Faustvii/StartupManager/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/Faustvii/StartupManager/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/Faustvii/StartupManager/releases/tag/v1.0.0